+++ mapchete_hub/__init__.py
@@ -2,7 +2,7 @@
 from ._core import mapchete_execute, mapchete_index, cleanup_config
 from ._misc import cleanup_datetime
 
-__version__ = "0.3"
+__version__ = "0.4"
 
 
 __all__ = [